+++ cmd/exec/main.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"os"
 
-	"github.com/deislabs/porter/pkg/mixin/exec"
+	"github.com/deislabs/porter/pkg/exec"
 	"github.com/spf13/cobra"
 )
