+++ cmd/exec/upgrade.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/deislabs/porter/pkg/mixin/exec"
+	"github.com/deislabs/porter/pkg/exec"
 	"github.com/spf13/cobra"
 )
