+++ pkg/exec/uninstall.go
